Split view.rs into pieces (refactor)
